test: remove duplicate test flows that are already in authorize.spec.ts
